Define NodeSpec, ParachainNodeSpec
Also moved some more primitives to common
