+++ cumulus/polkadot-parachain/src/common/parachain/mod.rs
@@ -14,7 +14,10 @@
 // You should have received a copy of the GNU General Public License
 // along with Cumulus.  If not, see <http://www.gnu.org/licenses/>.
 
-use crate::common::{BuildImportQueue, BuildRpcExtensions, ConstructNodeRuntimeApi};
+use crate::common::{
+	command::NodeCommandRunner, BuildImportQueue as BuildImportQueueT, BuildRpcExtensions,
+	BuildSelectChain, ConstructNodeRuntimeApi, NodeBackend, NodeSpec, NodeSpecProvider,
+};
 use cumulus_client_cli::CollatorOptions;
 use cumulus_client_consensus_common::ParachainBlockImport as TParachainBlockImport;
 use cumulus_client_service::{
@@ -23,20 +26,23 @@ use cumulus_client_service::{
 };
 use cumulus_primitives_core::ParaId;
 use cumulus_relay_chain_interface::{OverseerHandle, RelayChainInterface};
+use frame_support::DefaultNoBound;
 use futures::prelude::*;
 use parachains_common::{Block, Hash};
 use polkadot_primitives::CollatorPair;
 use prometheus_endpoint::Registry;
-use sc_consensus::{DefaultImportQueue, ImportQueue};
+use sc_consensus::ImportQueue;
 use sc_executor::WasmExecutor;
 use sc_network::{config::FullNetworkConfiguration, service::traits::NetworkBackend, NetworkBlock};
-use sc_service::{Configuration, PartialComponents, TFullBackend, TFullClient, TaskManager};
-use sc_telemetry::{Telemetry, TelemetryHandle, TelemetryWorker, TelemetryWorkerHandle};
+use sc_service::{Configuration, TFullBackend, TFullClient, TaskManager};
+use sc_sysinfo::HwBench;
+use sc_telemetry::TelemetryHandle;
 use sc_transaction_pool::FullPool;
 use sp_keystore::KeystorePtr;
-use std::{pin::Pin, sync::Arc, time::Duration};
+use std::{marker::PhantomData, pin::Pin, sync::Arc, time::Duration};
 
 pub mod aura;
+pub mod rpc;
 
 #[cfg(not(feature = "runtime-benchmarks"))]
 pub type HostFunctions = cumulus_client_service::ParachainHostFunctions;
@@ -54,16 +60,6 @@ pub type ParachainBackend = TFullBackend<Block>;
 pub type ParachainBlockImport<RuntimeApi> =
 	TParachainBlockImport<Block, Arc<ParachainClient<RuntimeApi>>, ParachainBackend>;
 
-/// Assembly of PartialComponents (enough to run chain ops subcommands)
-pub type Service<RuntimeApi> = PartialComponents<
-	ParachainClient<RuntimeApi>,
-	ParachainBackend,
-	(),
-	DefaultImportQueue<Block>,
-	FullPool<Block, ParachainClient<RuntimeApi>>,
-	(ParachainBlockImport<RuntimeApi>, Option<Telemetry>, Option<TelemetryWorkerHandle>),
->;
-
 /// Checks that the hardware meets the requirements and print a warning otherwise.
 fn warn_if_slow_hardware(hwbench: &sc_sysinfo::HwBench) {
 	// Polkadot para-chains should generally use these requirements to ensure that the relay-chain
@@ -75,6 +71,31 @@ fn warn_if_slow_hardware(hwbench: &sc_sysinfo::HwBench) {
 			err
 		);
 	}
+}
+
+pub struct BuildParachainSelectChain;
+
+impl BuildSelectChain<Block> for BuildParachainSelectChain {
+	type SelectChain = ();
+
+	fn build_select_chain(_backend: Arc<NodeBackend<Block>>) -> Self::SelectChain {
+		()
+	}
+}
+
+#[derive(DefaultNoBound)]
+pub struct ParachainNode<RuntimeApi, BuildImportQueue>(PhantomData<(RuntimeApi, BuildImportQueue)>);
+
+impl<RuntimeApi, BuildImportQueue> NodeSpec for ParachainNode<RuntimeApi, BuildImportQueue>
+where
+	RuntimeApi: ConstructNodeRuntimeApi<Block, ParachainClient<RuntimeApi>>,
+	BuildImportQueue: BuildImportQueueT<Block, ParachainClient<RuntimeApi>>,
+{
+	type Block = Block;
+	type RuntimeApi = RuntimeApi;
+	type HostFunctions = HostFunctions;
+	type BuildImportQueue = BuildImportQueue;
+	type BuildSelectChain = BuildParachainSelectChain;
 }
 
 pub trait StartConsensus<RuntimeApi>
@@ -99,82 +120,24 @@ where
 	) -> Result<(), sc_service::Error>;
 }
 
-pub trait NodeSpec {
+pub trait ParachainNodeSpec {
 	type RuntimeApi: ConstructNodeRuntimeApi<Block, ParachainClient<Self::RuntimeApi>>;
 
-	type BuildImportQueue: BuildImportQueue<Block, ParachainClient<Self::RuntimeApi>> + 'static;
+	type BuildImportQueue: BuildImportQueueT<
+		Block,
+		ParachainClient<Self::RuntimeApi>,
+		BlockImport = ParachainBlockImport<Self::RuntimeApi>,
+	>;
 
 	type BuildRpcExtensions: BuildRpcExtensions<
-			ParachainClient<Self::RuntimeApi>,
-			ParachainBackend,
-			FullPool<Block, ParachainClient<Self::RuntimeApi>>,
-		> + 'static;
+		Block,
+		ParachainClient<Self::RuntimeApi>,
+		ParachainBackend,
+	>;
 
-	type StartConsensus: StartConsensus<Self::RuntimeApi> + 'static;
+	type StartConsensus: StartConsensus<Self::RuntimeApi>;
 
 	const SYBIL_RESISTANCE: CollatorSybilResistance;
-
-	/// Starts a `ServiceBuilder` for a full service.
-	///
-	/// Use this macro if you don't actually need the full service, but just the builder in order to
-	/// be able to perform chain operations.
-	fn new_partial(config: &Configuration) -> sc_service::error::Result<Service<Self::RuntimeApi>> {
-		let telemetry = config
-			.telemetry_endpoints
-			.clone()
-			.filter(|x| !x.is_empty())
-			.map(|endpoints| -> Result<_, sc_telemetry::Error> {
-				let worker = TelemetryWorker::new(16)?;
-				let telemetry = worker.handle().new_telemetry(endpoints);
-				Ok((worker, telemetry))
-			})
-			.transpose()?;
-
-		let executor = sc_service::new_wasm_executor(config);
-
-		let (client, backend, keystore_container, task_manager) =
-			sc_service::new_full_parts_record_import::<Block, Self::RuntimeApi, _>(
-				config,
-				telemetry.as_ref().map(|(_, telemetry)| telemetry.handle()),
-				executor,
-				true,
-			)?;
-		let client = Arc::new(client);
-
-		let telemetry_worker_handle = telemetry.as_ref().map(|(worker, _)| worker.handle());
-		let telemetry = telemetry.map(|(worker, telemetry)| {
-			task_manager.spawn_handle().spawn("telemetry", None, worker.run());
-			telemetry
-		});
-
-		let transaction_pool = sc_transaction_pool::BasicPool::new_full(
-			config.transaction_pool.clone(),
-			config.role.is_authority().into(),
-			config.prometheus_registry(),
-			task_manager.spawn_essential_handle(),
-			client.clone(),
-		);
-
-		let block_import = ParachainBlockImport::new(client.clone(), backend.clone());
-		let import_queue = Self::BuildImportQueue::build_import_queue(
-			client.clone(),
-			Box::new(block_import.clone()),
-			config,
-			telemetry.as_ref().map(|telemetry| telemetry.handle()),
-			&task_manager,
-		)?;
-
-		Ok(PartialComponents {
-			client,
-			backend,
-			task_manager,
-			keystore_container,
-			select_chain: (),
-			import_queue,
-			transaction_pool,
-			other: (block_import, telemetry, telemetry_worker_handle),
-		})
-	}
 
 	/// Start a node with the given parachain spec.
 	///
@@ -184,7 +147,7 @@ pub trait NodeSpec {
 		polkadot_config: Configuration,
 		collator_options: CollatorOptions,
 		para_id: ParaId,
-		hwbench: Option<sc_sysinfo::HwBench>,
+		hwbench: Option<HwBench>,
 	) -> Pin<Box<dyn Future<Output = sc_service::error::Result<TaskManager>>>>
 	where
 		Net: NetworkBackend<Block, Hash>,
@@ -192,7 +155,9 @@ pub trait NodeSpec {
 		Box::pin(async move {
 			let parachain_config = prepare_node_config(parachain_config);
 
-			let params = Self::new_partial(&parachain_config)?;
+			let params = ParachainNode::<Self::RuntimeApi, Self::BuildImportQueue>::new_partial(
+				&parachain_config,
+			)?;
 			let (block_import, mut telemetry, telemetry_worker_handle) = params.other;
 
 			let client = params.client.clone();
@@ -327,5 +292,56 @@ pub trait NodeSpec {
 
 			Ok(task_manager)
 		})
+	}
+}
+
+impl<T> NodeSpecProvider for T
+where
+	T: ParachainNodeSpec,
+{
+	type NodeSpec = ParachainNode<T::RuntimeApi, T::BuildImportQueue>;
+}
+
+pub trait DynParachainNodeSpec: NodeCommandRunner {
+	fn start_node(
+		self: Box<Self>,
+		parachain_config: Configuration,
+		polkadot_config: Configuration,
+		collator_options: CollatorOptions,
+		para_id: ParaId,
+		hwbench: Option<HwBench>,
+	) -> Pin<Box<dyn Future<Output = sc_service::error::Result<TaskManager>>>>;
+}
+
+impl<T> DynParachainNodeSpec for T
+where
+	T: NodeCommandRunner + ParachainNodeSpec,
+{
+	fn start_node(
+		self: Box<Self>,
+		parachain_config: Configuration,
+		polkadot_config: Configuration,
+		collator_options: CollatorOptions,
+		para_id: ParaId,
+		hwbench: Option<HwBench>,
+	) -> Pin<Box<dyn Future<Output = sc_service::error::Result<TaskManager>>>> {
+		match parachain_config.network.network_backend {
+			sc_network::config::NetworkBackendType::Libp2p =>
+				<Self as ParachainNodeSpec>::start_node::<sc_network::NetworkWorker<_, _>>(
+					parachain_config,
+					polkadot_config,
+					collator_options,
+					para_id,
+					hwbench,
+				),
+			sc_network::config::NetworkBackendType::Litep2p =>
+				<Self as ParachainNodeSpec>::start_node::<sc_network::Litep2pNetworkBackend>(
+					parachain_config,
+					polkadot_config,
+					collator_options,
+					para_id,
+					hwbench,
+				),
+		}
 	}
 }
