+++ cumulus/polkadot-parachain/src/main.rs
@@ -47,7 +47,6 @@ mod cli;
 mod command;
 mod common;
 mod fake_runtime_api;
-mod rpc;
 mod service;
 
 fn main() -> sc_cli::Result<()> {
